refactor selector mod
