+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "open-coroutine-timer",
     "open-coroutine-queue",
-    "open-coroutine-utils",
     "open-coroutine-core",
     "open-coroutine-hooks",
     "open-coroutine-macros",
