+++ open-coroutine-core/Cargo.toml
@@ -35,6 +35,12 @@ windows-sys = { version = "0.48.0", features = ["Win32_System_SystemInformation"
 [dev-dependencies]
 backtrace = "0.3.67"
 
+[build-dependencies]
+cfg-if = "1.0.0"
+
+[target.'cfg(target_os = "linux")'.build-dependencies]
+cc = "1.0.83"
+
 [features]
 default = ["full"]
