--- open-coroutine-utils/README.md
@@ -1,3 +1,0 @@
-# open-coroutine-utils
-
-## This crate provide version util support
